Fixed a bug in which autocrack could not be enabled using the CLI.
When autocrack is enabled using the CLI, eaphammer creates a FIFO and changes the 'use_autocrack=0' line in hostapd's config file
to 'use_autocrack=1'. This tells hostapd that it should write all captured hashes to the FIFO created by eaphammer. EAPHammer then runs an event loop that listens for incoming hashes
from hostapd sent through the FIFO, and attempts to crack any that it receives. Cracked hashes are appended to db/eaphammer.eap_user, which is used by hostapd to store credentials.

The problem was that eaphammer was attempting to write the value "True" to hostapd's config file, rather than 1, which was causing hostapd to ignore the FIFO. This bug caused two side
effects: the first was that hostapd was not writing hashes to the FIFO, and the second was that eaphammer's autocrack event loop would block, since when Python's open function is use
to open a FIFO it will block until the writer has opened it first.

The bug was fixed by modifying core/hostapd_config.py to convert the appropriate boolean values to numerical values, and then later to strings before being written to hostapd's conf file.